FISH-6034 Update api package name.
By default OSGi won't provide an 'import' or 'uses' block
for the API to the impl if they share the same package name,
which leads to the OSGi bundle classloader not knowing how
to load the impl. Having them in separate packages is a quick fix.
